// src/lib/api/model/eventMessagePartUpdated.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * opencode
 * opencode api
 * OpenAPI spec version: 0.0.3
 */
import type { EventMessagePartUpdatedProperties } from './eventMessagePartUpdatedProperties'

export interface EventMessagePartUpdated {
  type: 'message.part.updated'
  properties: EventMessagePartUpdatedProperties
}