// src/lib/api/model/eventMessageUpdated.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * opencode
 * opencode api
 * OpenAPI spec version: 0.0.3
 */
import type { EventMessageUpdatedProperties } from './eventMessageUpdatedProperties'

export interface EventMessageUpdated {
  type: 'message.updated'
  properties: EventMessageUpdatedProperties
}