// src/lib/api/model/eventSessionCreated.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * opencode
 * opencode api
 * OpenAPI spec version: 0.0.3
 */
import type { EventSessionCreatedProperties } from './eventSessionCreatedProperties'

export interface EventSessionCreated {
  type: 'session.created'
  properties: EventSessionCreatedProperties
}