// src/lib/api/model/eventSessionUpdated.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * opencode
 * opencode api
 * OpenAPI spec version: 0.0.3
 */
import type { EventSessionUpdatedProperties } from './eventSessionUpdatedProperties'

export interface EventSessionUpdated {
  type: 'session.updated'
  properties: EventSessionUpdatedProperties
}