// src/lib/api/model/textPart.ts
/**
 * Generated by orval v7.15.0 🍺
 * Do not edit manually.
 * opencode
 * opencode api
 * OpenAPI spec version: 0.0.3
 */
import type { TextPartTime } from './textPartTime'
import type { TextPartMetadata } from './textPartMetadata'

export interface TextPart {
  id: string
  sessionID: string
  messageID: string
  type: 'text'
  text: string
  synthetic?: boolean
  time?: TextPartTime
  metadata?: TextPartMetadata
}